fix(executor): preserve the process environment on Windows
Always populating cmd.Env from buildEnv stripped Windows-specific
variables (SystemRoot, PSModulePath, USERPROFILE, APPDATA, TEMP,
ProgramData, etc.) that PowerShell and other Windows processes require;
their absence breaks process startup. Privilege demotion is a no-op on
Windows, so there is no service-env leakage concern there.

Introduce a platform-specific processBaseEnv: empty on Unix (deterministic,
no inheritance) and the inherited process environment on Windows, mirroring
the Websh path in pty_windows.go. Move the Unix-only no-leak test under a
build tag.

Addresses Copilot review on #322.
